Triage Ruff 0.16 default-ruleset errors and lift the CI pin (#729)
Switch pyproject.toml to an explicit `select` (was `extend-select` over
Ruff's defaults) so a future Ruff release can't silently widen the
ruleset under CI again. Adopt the new-default rule categories worth
having (RUF*, SIM*, C4*, PERF*, PIE*, EXE001, PLW1510, PLR*, PLC3002,
G201, ISC004, S110/S112), and deliberately leave out BLE and DTZ with
documented rationale — both conflict with intentional patterns already
in this codebase (catch-all error reporting in the worker/task-runner,
and naive-UTC DateTime columns).

Resolves all 454 errors Ruff 0.16.1 reports against the pre-existing
tree: ~421 mechanically via `ruff check --fix`/`--unsafe-fixes`
(unused-unpacked-variable renames, stale noqa removal, dict/list
literals, etc.), the remaining 33 by hand (chmod +x for EXE001,
explicit subprocess.run(check=False), ClassVar annotations, collapsed
nested if/with, debug logging for previously-silent except blocks,
logging.exception, and a couple of small rewrites).

Bump the CI ruff pin to 0.16.1 now that the ruleset is triaged.

+++ myapp/services/similarity_stats.py
@@ -43,7 +43,7 @@ def _percentile(sorted_vals, p):
         return 0
     if p >= 100:
         return sorted_vals[-1]
-    k = max(0, min(len(sorted_vals) - 1, int(round((p / 100.0) * len(sorted_vals) + 0.5)) - 1))
+    k = max(0, min(len(sorted_vals) - 1, round((p / 100.0) * len(sorted_vals) + 0.5) - 1))
     return sorted_vals[k]
